+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 
 #### Bugfixes
 
-* fix bug that caused control state value to not be properly set to form element if the ID of the control state changed but the state's value was the same as the last value the view reported for the previous state
+* fix issue that caused control state value to not be properly set to form element if the ID of the control state changed but the state's value was the same as the last value the view reported for the previous state
+* changed form control state directive to run its initialization code inside the `ngAfterViewInit` hook instead of `ngOnInit` to allow proper interaction with form elements that can have dynamically rendered children they depend on (e.g. dynamic `option`s for `select`s)
 
 <a name="1.0.0"></a>
 ### 1.0.0